+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 0.1.2
+## 0.1.3+13
+
+- Added `skydroid validate` command
+- Fixed default version regex for native Android apps
+
+## 0.1.2+12
 
 - Added Split-APK support